[FIX] hr_personal_equipment_stock: Broken receptions with dropshipping
When the stock_dropshipping module was installed in addition to
hr_personal_equipment_stock, purchase orders for fulfilling personal
equipment requests would result in a Dropship picking instead of a
standard reception picking being created. This makes it such that the
products were received at the wrong location (Customers location).

This was caused by the location_dest_id being set on the purchase order
created when the procurement group's run() method is called.

The purchase order no longer has a location_dest_id set, making it
correctly create a receipt that sends the goods to the location
specified on the personal equipment request.